i committed the MyFirebaseDatabase.java file
Within this file, I added the method required to create a node in the firebase that has all the event categories and add event categories. the method takes as input a string which represents the event category, and puts it into the firebase. But I haven't yet added a method in the admin class that uses this method to actually do the action of adding an event category.