I am committing the files Admin.java and MyFirebaseDatabase.java
For Admin.java, i created the method addEventCategory which takes in as an argument the name of the event category as a string, and uses the createEventCategory method in the MyFirebaseDatabase class to be able to add it to the database. 
For MyFirebaseDatabase.java, i created the method getAllChildNodes which gets all child nodes of a node in the database. for now, the method isn't done yet, its return type is void since the method doesn't yet store all the values in a list to return them. also i need to make the method more genera, currently it just returns key as well as a description field. but i need to try to generalize ths to be usable on many different types of nodes that have different fields. i also fixed bugs in MyFirebaseDatabase.