commiting 4 files
in the AddCategoryActivity, all i did was i added some comments to better understand the code. in the User.java class, i added an attribute id and i updated both constructors with arguments, so that they take in String id as an argument and add it to the user type object. in RegisterActivity, i updated the code so that the hashmap which stores all the user information also stores the user's firebase id as the field id. and all that was foe me to be able to do stuff in UserAdapter. within that class, i made the btnEdit and btnDelete, and i worked on the code btnDelete so that it takes that specific user, which is found by their user id, and deletes them. what's missing for this code to be functional is that i need to make sure that when the user type objects are instantiated and put in the ArrayList, that they are correctly instantiated with their id, so that user.id works. also a second thing is that often we want a confirmation message asking the admin, "do you really want to delete this user?". I have not implemented this.